Create NewsFeedController
